requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
pymongo==4.3.3
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==2.3.7
scikit-learn==1.3.0
nltk==3.8.1
